Add Message struct and implementation
Add a Message struct representing a single rpc message that has
a field for the called method and the data as well as the length
indicator and a crc32 checksum. The struct implements a way of
serializing the message into bytes and constructing it from bytes
with support for serde.

Signed-off-by: trivernis <trivernis@protonmail.com>